Controller and object adjustments around package slots
