+++ LBH.AdultSocialCare.Api/GlobalSuppressions.cs
@@ -8,3 +8,4 @@ using System.Diagnostics.CodeAnalysis;
 [assembly: SuppressMessage("Design", "CA1031:Do not catch general exception types", Justification = "<Pending>", Scope = "member", Target = "~M:LBH.AdultSocialCare.Api.V1.Controllers.ClientController.Create(LBH.AdultSocialCare.Api.V1.Boundary.Request.ClientsRequest)~System.Threading.Tasks.Task{Microsoft.AspNetCore.Mvc.ActionResult{LBH.AdultSocialCare.Api.V1.Boundary.Response.ClientsResponse}}")]
 [assembly: SuppressMessage("Design", "CA1031:Do not catch general exception types", Justification = "<Pending>", Scope = "member", Target = "~M:LBH.AdultSocialCare.Api.V1.Controllers.HomeCarePackageController.Create(LBH.AdultSocialCare.Api.V1.Boundary.Request.HomeCarePackageRequest)~System.Threading.Tasks.Task{Microsoft.AspNetCore.Mvc.ActionResult{LBH.AdultSocialCare.Api.V1.Boundary.Response.HomeCarePackageResponse}}")]
 [assembly: SuppressMessage("Design", "CA1031:Do not catch general exception types", Justification = "<Pending>", Scope = "member", Target = "~M:LBH.AdultSocialCare.Api.V1.Controllers.HomeCarePackageController.Create(LBH.AdultSocialCare.Api.V1.Boundary.Request.HomeCare.HomeCarePackageRequest)~System.Threading.Tasks.Task{Microsoft.AspNetCore.Mvc.ActionResult{LBH.AdultSocialCare.Api.V1.Boundary.Response.HomeCarePackageResponse}}")]
+[assembly: SuppressMessage("Design", "CA1031:Do not catch general exception types", Justification = "<Pending>", Scope = "member", Target = "~M:LBH.AdultSocialCare.Api.V1.Controllers.HomeCare.HomeCarePackageController.Create(LBH.AdultSocialCare.Api.V1.Boundary.Request.HomeCare.HomeCarePackageRequest)~System.Threading.Tasks.Task{Microsoft.AspNetCore.Mvc.ActionResult{LBH.AdultSocialCare.Api.V1.Boundary.Response.HomeCarePackageResponse}}")]
