+++ LBH.AdultSocialCare.Api/V1/Constants.cs
@@ -1,7 +1,11 @@
 namespace LBH.AdultSocialCare.Api.V1
 {
+
     public static class Constants
     {
+
         public const string CorrelationId = "x-correlation-id";
+
     }
+
 }
