+++ LBH.AdultSocialCare.Api/V1/Controllers/HomeCare/HomeCarePackageSlotsController.cs
@@ -1,32 +1,33 @@
-using LBH.AdultSocialCare.Api.V1.Boundary.Request;
+using System;
+using System.Threading.Tasks;
+using LBH.AdultSocialCare.Api.V1.Boundary.Request.HomeCare;
 using LBH.AdultSocialCare.Api.V1.Boundary.Response;
 using LBH.AdultSocialCare.Api.V1.Domain;
 using LBH.AdultSocialCare.Api.V1.Factories;
 using LBH.AdultSocialCare.Api.V1.UseCase.Interfaces;
 using Microsoft.AspNetCore.Mvc;
-using System;
-using System.Threading.Tasks;
-using LBH.AdultSocialCare.Api.V1.Boundary.Request.HomeCare;
 
-namespace LBH.AdultSocialCare.Api.V1.Controllers
+namespace LBH.AdultSocialCare.Api.V1.Controllers.HomeCare
 {
 
-    [Route("api/v1/homeCarePackageSlots")]
+    [Route("api/v1/[controller]")]
     [Produces("application/json")]
     [ApiController]
     public class HomeCarePackageSlotsController : BaseController
     {
 
         private readonly IUpsertHomeCarePackageSlotsUseCase _upsertHomeCarePackageSlotsUseCase;
-        private readonly IDeleteHomeCarePackageSlotsUseCase _deleteHomeCarePackageSlotsUseCase;
 
-        public HomeCarePackageSlotsController(IUpsertHomeCarePackageSlotsUseCase upsertHomeCarePackageSlotsUseCase,
-            IDeleteHomeCarePackageSlotsUseCase deleteHomeCarePackageSlotsUseCase)
+        public HomeCarePackageSlotsController(IUpsertHomeCarePackageSlotsUseCase upsertHomeCarePackageSlotsUseCase)
         {
             _upsertHomeCarePackageSlotsUseCase = upsertHomeCarePackageSlotsUseCase;
-            _deleteHomeCarePackageSlotsUseCase = deleteHomeCarePackageSlotsUseCase;
         }
 
+        /// <summary>
+        /// Creates the specified home care package slots.
+        /// </summary>
+        /// <param name="homeCarePackageSlotsRequestList">The home care package slots request list.</param>
+        /// <returns>A home care package slots response list.</returns>
         [HttpPost]
         public async Task<ActionResult<HomeCarePackageSlotsResponseList>> Create(
             HomeCarePackageSlotsRequestList homeCarePackageSlotsRequestList)
@@ -41,7 +42,10 @@ namespace LBH.AdultSocialCare.Api.V1.Controllers
                         .ExecuteAsync(homeCarePackageSlotListDomain)
                         .ConfigureAwait(false));
 
-                if (homeCarePackageSlotsResponse == null) return NotFound();
+                if (homeCarePackageSlotsResponse == null)
+                {
+                    return NotFound();
+                }
 
                 return Ok(homeCarePackageSlotsResponse);
             }
