+++ LBH.AdultSocialCare.Api/V1/CorrelationMiddleware.cs
@@ -5,8 +5,10 @@ using Microsoft.AspNetCore.Http;
 
 namespace LBH.AdultSocialCare.Api.V1
 {
+
     public class CorrelationMiddleware
     {
+
         private readonly RequestDelegate _next;
 
         public CorrelationMiddleware(RequestDelegate next)
@@ -24,14 +26,17 @@ namespace LBH.AdultSocialCare.Api.V1
             if (_next != null)
                 await _next(context).ConfigureAwait(false);
         }
+
     }
 
     public static class CorrelationMiddlewareExtensions
     {
-        public static IApplicationBuilder UseCorrelation(
-            this IApplicationBuilder builder)
+
+        public static IApplicationBuilder UseCorrelation(this IApplicationBuilder builder)
         {
             return builder.UseMiddleware<CorrelationMiddleware>();
         }
+
     }
+
 }
