+++ LBH.AdultSocialCare.Api/V1/UseCase/HomeCare/GetAllHomeCareServiceTypesUseCase.cs
@@ -1,11 +1,10 @@
-using LBH.AdultSocialCare.Api.V1.Gateways.Interfaces;
-using LBH.AdultSocialCare.Api.V1.Infrastructure.Entities;
-using LBH.AdultSocialCare.Api.V1.UseCase.Interfaces;
 using System.Collections.Generic;
 using System.Threading.Tasks;
+using LBH.AdultSocialCare.Api.V1.Gateways.Interfaces;
 using LBH.AdultSocialCare.Api.V1.Infrastructure.Entities.HomeCare;
+using LBH.AdultSocialCare.Api.V1.UseCase.Interfaces;
 
-namespace LBH.AdultSocialCare.Api.V1.UseCase
+namespace LBH.AdultSocialCare.Api.V1.UseCase.HomeCare
 {
     public class GetAllHomeCareServiceTypesUseCase : IGetAllHomeCareServiceTypesUseCase
     {
