+++ LBH.AdultSocialCare.Api/V1/UseCase/HomeCare/GetTimeSlotTypesUseCase.cs
@@ -1,11 +1,11 @@
+using System;
+using System.Threading.Tasks;
 using LBH.AdultSocialCare.Api.V1.Domain;
 using LBH.AdultSocialCare.Api.V1.Factories;
 using LBH.AdultSocialCare.Api.V1.Gateways.Interfaces;
 using LBH.AdultSocialCare.Api.V1.UseCase.Interfaces;
-using System;
-using System.Threading.Tasks;
 
-namespace LBH.AdultSocialCare.Api.V1.UseCase
+namespace LBH.AdultSocialCare.Api.V1.UseCase.HomeCare
 {
     public class GetTimeSlotTypesUseCase : IGetTimeSlotTypesUseCase
     {
