+++ LBH.AdultSocialCare.Api/V1/UseCase/HomeCare/UpsertHomeCarePackageSlotsUseCase.cs
@@ -1,9 +1,9 @@
+using System.Threading.Tasks;
 using LBH.AdultSocialCare.Api.V1.Domain;
 using LBH.AdultSocialCare.Api.V1.Gateways.Interfaces;
 using LBH.AdultSocialCare.Api.V1.UseCase.Interfaces;
-using System.Threading.Tasks;
 
-namespace LBH.AdultSocialCare.Api.V1.UseCase
+namespace LBH.AdultSocialCare.Api.V1.UseCase.HomeCare
 {
 
     public class UpsertHomeCarePackageSlotsUseCase : IUpsertHomeCarePackageSlotsUseCase
