+++ LBH.AdultSocialCare.Api/V1/UseCase/HomeCare/UpsertHomeCarePackageUseCase.cs
@@ -1,10 +1,10 @@
+using System.Threading.Tasks;
 using LBH.AdultSocialCare.Api.V1.Domain;
 using LBH.AdultSocialCare.Api.V1.Factories;
 using LBH.AdultSocialCare.Api.V1.Gateways.Interfaces;
 using LBH.AdultSocialCare.Api.V1.UseCase.Interfaces;
-using System.Threading.Tasks;
 
-namespace LBH.AdultSocialCare.Api.V1.UseCase
+namespace LBH.AdultSocialCare.Api.V1.UseCase.HomeCare
 {
     public class UpsertHomeCarePackageUseCase : IUpsertHomeCarePackageUseCase
     {
